remove dtype from np.stack

+++ src/aspire/abinitio/commonline_sdp.py
@@ -188,7 +188,7 @@ class CommonlineSDP(CLOrient3D):
         r1 = np.dot(A.T, v1)
         r2 = np.dot(A.T, v2)
         r3 = np.cross(r1, r2, axis=0)
-        rotations = np.stack((r1.T, r2.T, r3.T), axis=-1, dtype=self.dtype)
+        rotations = np.stack((r1.T, r2.T, r3.T), axis=-1)
 
         # Make sure that we got rotations by enforcing R to be
         # a rotation (in case the error is large)
